Change type of parameter to match type in definition

+++ src/Env.h
@@ -45,7 +45,7 @@ struct Parameters {
     double new_erc_name_probability;// = 0.001
     int max_points_in_random_expression;// = 50
     int max_points_in_program;// = 100
-    int random_seed;// = time.time()
+    long random_seed;// = time.time()
 };
 
 extern const Parameters global_parameters;
